split into smaller components
